lambda/variants/requirements.txt
boto3==1.34.51
Pillow==10.2.0
